refactor(constants): make enums into top level file
