+++ src/monopoly/enums.py
